% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCloudAnalysis.R
\name{chiMaps}
\alias{chiMaps}
\title{Chi maps of two variables}
\usage{
chiMaps(stack1, stack2)
}
\arguments{
\item{stack1}{\code{rasterStack}}

\item{stack2}{\code{rasterStack} containing the same number of \code{rasterLayer}s and has euqal extent and resolution.}
}
\value{
A \code{rasterStack} containing the chi maps.
}
\description{
Calculates the chi maps for all the raster pairs stored in the two \code{rasterStack}s.
As observed values, the first stack is used. The second stack is set to be the expected value.
}
\examples{
chiMaps(stack1, stack2)
}
