% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData3D.R
\docType{data}
\name{dem}
\alias{dem}
\title{Example digital elevation model (DEM)}
\description{
This is data to be included in the package and can be used to test its functionality.
The \code{dem} data is a \code{RasterLayer} and has a resolution of 90 meters. It is the topography
of the Swiss midlands. The complete dataset can be downloaded directly from http://srtm.csi.cgiar.org/srtmdata/.
}
\references{
\url{http://srtm.csi.cgiar.org/srtmdata/}
}
\keyword{data}
