% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{dem2track.extent}
\alias{dem2track.extent}
\title{Crops the DEM to the extent of the track with a buffer}
\usage{
dem2track.extent(DEM, track, buffer = 100)
}
\arguments{
\item{DEM}{a raster containing a digital elevation model, covering the extent as the track}

\item{track}{data.frame with x,y,z coordinates of the original track}

\item{buffer}{buffer with, by default set to 100}
}
\value{
A the cropped digital elevation model as a raster layer.
}
\description{
Crops the DEM to the extent of the track with a buffer
}
\examples{
dem2track.extent(dem, niclas)
}
