% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{dist2target.3d}
\alias{dist2target.3d}
\title{Distance to target}
\usage{
dist2target.3d(track)
}
\arguments{
\item{track}{a track data.frame containing x, y and z coordinates}
}
\value{
A numeric vector with the distances to target
}
\description{
Calculates the distance between every point in the track and the last point (target).
}
\examples{
dist2target.3d(niclas)
}
