% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{.check.extent}
\alias{.check.extent}
\title{Checks if the track lies inside the digital elevation model.}
\usage{
.check.extent(DEM, track)
}
\arguments{
\item{DEM}{a 'RasterLayer' containing a digital elevation model}

\item{track}{a data.frame with 3 columns containing the x,y,z coordinates}
}
\value{
A logical: TRUE if the track lies inside the DEM, FALSE otherwise.
}
\description{
Checks if the track lies inside the digital elevation model.
}
\examples{
.check.extent(DEM, track)
}
