% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core3D.R
\name{.cutMidpoints}
\alias{.cutMidpoints}
\title{Find corresponding midpoints of breaks}
\usage{
.cutMidpoints(x, breaks, rm.empty = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{breaks}{integer, which defines the number of breaks}

\item{rm.empty}{logical: should not occuring levels in the factor be removed?}
}
\value{
Returns a factor containing the midpoint
}
\description{
Assigns each element of a input vector the mid point of the corresponding interval.
Adjusted the original \code{cut()} function from the base package.
The midpoints are turned to a factor which facilitates the further histogram creation.
}
\examples{
.cutMidpoints(x, breaks)
}
