% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfWrapper3D.R
\name{.df2sf.3d}
\alias{.df2sf.3d}
\title{Converts a track data.frame to a \code{'sf, data.frame'}}
\usage{
.df2sf.3d(track, CRS = NA)
}
\arguments{
\item{track}{eRTG3D track data.frame}

\item{CRS}{string containing the proj4 code of the CRS}
}
\value{
A track of type \code{'sf, data.frame'}.
}
\description{
Converts a track data.frame to a \code{'sf, data.frame'}
}
\examples{
.df2sf.3d(track, CRS = NA)
}
