% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{.distance.2d}
\alias{.distance.2d}
\title{Ground distance covered in XY plane}
\usage{
.distance.2d(dx, dy)
}
\arguments{
\item{dx}{distance in x}

\item{dy}{distance in y}
}
\value{
Ground distance in XY plane
}
\description{
Ground distance covered in XY plane
}
\examples{
.distance.2d(dx, dy)
}
