% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{.distance.3d}
\alias{.distance.3d}
\title{Distance covered in 3 dimensions, radius}
\usage{
.distance.3d(dx, dy, dz)
}
\arguments{
\item{dx}{distance in x}

\item{dy}{distamce in y}

\item{dz}{distance in z}
}
\value{
The radius
}
\description{
Distance covered in 3 dimensions, radius
}
\examples{
.distance.3d(dx, dy, dz)
}
