% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core3D.R
\name{.fd.bw}
\alias{.fd.bw}
\title{Freedman–Diaconis rule}
\usage{
.fd.bw(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
The bandwith
}
\description{
In statistics, this rule can be used to select the size
of the bins to be used in a histogram.
}
\examples{
.df.bw(x)
}
