% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{.get.azimut}
\alias{.get.azimut}
\title{Azimut in respect to the x-axis}
\usage{
.get.azimut(dx, dy)
}
\arguments{
\item{dx}{distance in x}

\item{dy}{distance in y}
}
\value{
The azimut in radian
}
\description{
Azimut in respect to the x-axis
}
\examples{
.get.azimut(dx, dy)
}
