% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{.get.polar}
\alias{.get.polar}
\title{Polar angle/gradient in respect to the z-axis}
\usage{
.get.polar(d, dz)
}
\arguments{
\item{d}{ground distance in xy plane}

\item{dz}{distance in z}
}
\value{
The polar angle in radian
}
\description{
Polar angle/gradient in respect to the z-axis
}
\examples{
.get.polar(d, dz)
}
