% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{.is.df.xyz}
\alias{.is.df.xyz}
\title{Tests if the object is of type 'data.frame' and has x, y, z coordinates without NA values}
\usage{
.is.df.xyz(track)
}
\arguments{
\item{track}{any object to test}
}
\value{
A logical: TRUE if the track is the object needed, FALSE otherwise.
}
\description{
Tests if the object is of type 'data.frame' and has x, y, z coordinates without NA values
}
\examples{
.is.df.xyz(track)
}
