% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfWrapper3D.R
\name{.matrix2sf.3d}
\alias{.matrix2sf.3d}
\title{Converts a track matrix to a sf data.frame}
\usage{
.matrix2sf.3d(track, CRS = NA)
}
\arguments{
\item{track}{matrix with x, y and z coordinates.}

\item{CRS}{string containing the proj4 code of the CRS}
}
\value{
A track of type \code{'sf, data.frame'}.
}
\description{
Converts a track matrix to a sf data.frame
}
\examples{
.matrix2sf.3d(track, CRS = NA)
}
