% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfWrapper3D.R
\name{.move2sf.3d}
\alias{.move2sf.3d}
\title{Converts a move object to a sf data.frame}
\usage{
.move2sf.3d(track)
}
\arguments{
\item{track}{move object with a single track}
}
\value{
A track of type \code{'sf, data.frame'}.
}
\description{
Converts a move object to a sf data.frame
}
\examples{
.move2sf.3d(track)
}
