% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{.plot3d.density}
\alias{.plot3d.density}
\title{Density plot of one or two variables}
\usage{
.plot3d.density(values1, values2 = NULL, titleText = character(1),
  xlab = "x", ylab = "P(x)", scaleDensity = FALSE)
}
\arguments{
\item{values1}{numeric vector}

\item{values2}{numeric vector or \code{NULL}}

\item{titleText}{character of the title}

\item{xlab}{character of x label}

\item{ylab}{character of y label}

\item{scaleDensity}{logical: should density be scaled between 0 and 1, then sum of the area under the curve is not 1 anymore!}
}
\value{
A ggplot2 object.
}
\description{
Density plot of one or two variables
}
\examples{
plot3d.density(values)
}
