% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelUnix3D.R
\name{.qProb.3d.unix}
\alias{.qProb.3d.unix}
\title{Parallel or single core Q probabilities for n steps on unix (and windows single core)}
\usage{
.qProb.3d.unix(sim, n.locs, maxBin = 25)
}
\arguments{
\item{sim}{the result of simm.uncond.3d(), or a data frame with at least
x,y,z-coordinates, the arrival azimuth and the arrival gradient.}

\item{n.locs}{number of total segments to be modelled,
the length of the desired conditional empirical random walk}

\item{maxBin}{numeric scalar, maximum number of bins per dimension of the tld-cube (\link[eRTG3D]{turnLiftStepHist})}
}
\value{
A list containing the Q - tldCubes for every step
}
\description{
Calculates the Q probability, representing the pull to
the target. The number of steps on which the Q prob will be
quantified is number of total segments less than one
(the last step is defined by the target itself).
}
\examples{
.qProb.3d.unix(sim, n.locs)
}
