% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing3D.R
\name{.test2text}
\alias{.test2text}
\title{Extract test results as string}
\usage{
.test2text(test, alpha)
}
\arguments{
\item{test}{object of type \code{htest}}

\item{alpha}{scalar: significance level, default \code{alpha = 0.05}}
}
\value{
A character describing the results.
}
\description{
Extract test results as string
}
\examples{
.test2text(test, alpha)
}
