% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core3D.R
\name{.wrap}
\alias{.wrap}
\title{Wraps angles to [0,2pi] degrees}
\usage{
.wrap(x)
}
\arguments{
\item{x}{the angle in radians}
}
\value{
The wrapped angle
}
\description{
Wraps angles to [0,2pi] degrees
}
\examples{
.wrap(x)
}
