% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core3D.R
\name{get.densities.3d}
\alias{get.densities.3d}
\title{Extract tldCube and autodifferences functions from a track}
\usage{
get.densities.3d(track, heightDistEllipsoid = TRUE, DEM = NULL)
}
\arguments{
\item{track}{a data.frame with 3 columns containing the x,y,z coordinates}

\item{heightDistEllipsoid}{logical: Should a distribution of the flight height over ellipsoid be extracted and later used in the sim.cond.3d()?}

\item{DEM}{a raster containting a digital elevation model, covering the same extent as the track}
}
\value{
A list containing the tldCube and the autodifferences functions (and additionally the height distribution function)
}
\description{
Get densities creates a list consisting of the 3 dimensional
probability distribution cube for turning angle, lift angle and step length
as well as the uni-dimensional distributions of the differences
of the turning angles, lift angles and step lengths with a lag of 1 to maintain
minimal level of autocorrelation in each of the terms.
}
\examples{
get.densities.3d(track, heightDist = TRUE)
}
