% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glidingSoaring3D.R
\name{get.glideRatio.3d}
\alias{get.glideRatio.3d}
\title{Calculate glide ratio}
\usage{
get.glideRatio.3d(track)
}
\arguments{
\item{track}{a track data.frame containing x, y and z coordinates of a gliding section}
}
\value{
The ratio between horizontal and vertical movement.
}
\description{
Calculates the ratio between horizontal movement and vertical movement.
The value expresses the distance covered forward movement per distance movement in sinking.
}
\examples{
get.glideRatio.3d(niclas)
}
