% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfWrapper.R
\name{is.sf}
\alias{is.sf}
\title{Test if the object is of type 'sf, data.frame'}
\usage{
is.sf(track)
}
\arguments{
\item{track}{any object to test}
}
\value{
A logical: TRUE if the track is the sf object needed, FALSE otherwise.
}
\description{
Test if the object is of type 'sf, data.frame'
}
\examples{
is.sf(track)
}
