% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{lift2target.3d}
\alias{lift2target.3d}
\title{Lift angle to target}
\usage{
lift2target.3d(track)
}
\arguments{
\item{track}{a track data.frame containing x, y and z coordinates}
}
\value{
A numeric vector with the lift angles to target
}
\description{
Calculates the lift angle between every point in the track and the last point (target).
}
\examples{
lift2target.3d(niclas)
}
