% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCloudAnalysis3D.R
\name{logRasterStack}
\alias{logRasterStack}
\title{Converts a rasterStack to logarithmic scale}
\usage{
logRasterStack(rStack, standartize = FALSE, InfVal = NA)
}
\arguments{
\item{rStack}{rasterStack to convert to logarithmic scale}

\item{standartize}{logical: standartize cube between 0 and 1}

\item{InfVal}{the value that \code{Inf} and \code{-Inf} should be rpeplaced with}
}
\value{
A rasterStack in logarithmic scale
}
\description{
Avoids the problem of -Inf occuring for log(0).
}
\examples{
logRasterStack(raster::stack(dem))
}
