% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{movingMedian}
\alias{movingMedian}
\title{Applies a moving median over a vector}
\usage{
movingMedian(data, window)
}
\arguments{
\item{data}{numeric vector}

\item{window}{uneven number for the size of the moving window}
}
\value{
A numeric vector.
}
\description{
Applies a moving median over a vector
}
\examples{
movingMedian(data, window = 5)
}
