% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simModes3D.R
\name{n.sim.cond.modes.3d}
\alias{n.sim.cond.modes.3d}
\title{Multiple Conditioned Empirical Random Walks (CERW) with modes in 3D}
\usage{
n.sim.cond.modes.3d(n.sim, locsVec, start = c(0, 0, 0), end = start, a0, g0,
  dList, qList, glideRatio = 20, error = FALSE, multicore = FALSE, MODE,
  DEM = NULL, BG = NULL)
}
\arguments{
\item{n.sim}{number of CERWs to simulate}

\item{locsVec}{length of the trajectory in locations}

\item{start}{numeric vector of length 3 with the coordinates of the start point}

\item{end}{numeric vector of length 3 with the coordinates of the end point}

\item{a0}{initial incoming heading in radian}

\item{g0}{initial incoming gradient/polar angle in radian}

\item{error}{logical: add random noise to the turn angle, lift angle and step length to account for errors measurements?}

\item{multicore}{logical: run computations in parallel (n-1 cores)?}

\item{DEM}{raster layer containing a digital elevation model, covering the area between start and end point}

\item{BG}{a background raster layer that can be used to inform the choice of steps}

\item{densities}{list object returned by get.densities.3d() function}

\item{qProbs}{list object returned by qProb.3d() function}
}
\value{
A list containing the CERWs or NULLs if dead ends have been encountered.
}
\description{
Creates n conditioned empirical random walks using different modes, with a specific starting and ending point,
geometrically similar to the initial trajectory by applying \link[eRTG3D]{sim.cond.modes.3d} multiple times.
}
\examples{
n.sim.cond.modes.3d(n.sim, locsVec, start = c(0,0,0), end=start, a0, g0, dList, qList, MODE)
}
