% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData3D.R
\docType{data}
\name{niclas}
\alias{niclas}
\title{Example track data.frame}
\description{
This is data to be included in the package and can be used to test its functionality.
The track consists of x, y and z coordinates and represents the movement of a stork
called \code{niclas} in the Swiss midlands.
}
\references{
\url{https://www.movebank.org}
}
\keyword{data}
