% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plot2d}
\alias{plot2d}
\title{Plot function to plot the tracks in 2d}
\usage{
plot2d(origTrack, cerwList = NULL, titleText = character(1), DEM = NULL)
}
\arguments{
\item{origTrack}{a data.frame with x,y,z coordinates}

\item{cerwList}{a list containing a data.frame with x,y,z coordinates or a data.frame}

\item{titleText}{string with title of the plot}

\item{DEM}{an object of type 'RasterLayer', needs overlapping extent with the lines}
}
\value{
Nothing, plots a 2D ggplot2 object.
}
\description{
Plot function to plot the tracks in 2d
}
\examples{
plot2d(track)
}
