% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3d.multiplot}
\alias{plot3d.multiplot}
\title{Multiple plot function for ggplot objects}
\usage{
plot3d.multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{a list of ggplot objects}

\item{cols}{number of columns in layout}

\item{layout}{a matrix specifying the layout. If present, \code{cols} is ignored.}
}
\value{
Nothing, plots the ggplot2 objects.
}
\description{
If the layout is something like \code{matrix(c(1,2,3,3), nrow=2, byrow=TRUE)},
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\examples{
plot3d.multiplot(plot2d(niclas), plot2d(niclas), plot2d(niclas))
}
