% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3d.new}
\alias{plot3d.new}
\title{Plot 3D track(s) with a surface}
\usage{
plot3d.new(origTrack, simTrack = NULL, titleText = character(1),
  DEM = NULL, padding = 0.05, distanceHeightRatio = 10)
}
\arguments{
\item{origTrack}{a list containing data.frames with x,y,z coordinates or a data.frame}

\item{simTrack}{a list containing data.frames with x,y,z coordinates or a data.frame}

\item{titleText}{string with title of the plot}

\item{DEM}{an object of type 'RasterLayer', needs overlapping extent with the line(s)}

\item{padding}{adds a pad to the 2-D space in percentage (by default set to 0.05)}

\item{distanceHeightRatio}{ratio of the height in respect to the ground distance (by default ratio is 10)}
}
\value{
Plots a 3D plotly object
}
\description{
Plot 3D track(s) with a surface
}
\examples{
plot3d(track)
}
