% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3d}
\alias{plot3d}
\title{Plot 3D track(s) with a surface}
\usage{
plot3d(origTrack, cerwList = NULL, titleText = character(1), DEM = NULL,
  maxHeight = 8000)
}
\arguments{
\item{origTrack}{a data.frame with x,y,z coordinates}

\item{cerwList}{a list containing a data.frame with x,y,z coordinates or a data.frame}

\item{titleText}{string with title of the plot}

\item{DEM}{an object of type 'RasterLayer', needs overlapping extent with the lines}

\item{maxHeight}{Maximum plot height, default 8000m}
}
\value{
Plots a 2D ggplot2 object
}
\description{
Plot 3D track(s) with a surface
}
\examples{
plot3d(track)
}
