% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3d.tldCube}
\alias{plot3d.tldCube}
\title{Visualize turn-lift-step histogram}
\usage{
plot3d.tldCube(tldCube)
}
\arguments{
\item{tldCube}{tldCube; the ouptut from \link[eRTG3D]{turnLiftStepHist} or \link[eRTG3D]{get.densities.3d}}
}
\value{
Plots a plotly object
}
\description{
Creates a three dimensional scatterplot of the possibles next steps,
based on the tldCube, which was extracted from a track.
}
\examples{
P <- get.track.densities.3d(niclas)
suppressWarnings(plot3d.tldCube(P$tldCube))
}
