% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCloudAnalysis.R
\name{plotRaster}
\alias{plotRaster}
\title{Plots a rasterLayer or rasterStack}
\usage{
plotRaster(r, title = character(0), centerColorBar = FALSE)
}
\arguments{
\item{r}{\code{rasterLayer} or \code{rasterStack}}

\item{title}{title text of plot(s)}

\item{centerColorBar}{logical: center colobar around 0 and use \code{RdBuTheme()}?}
}
\value{
Plots the rasters
}
\description{
Plots a rasterLayer or rasterStack
}
\examples{
plotRaster(r)
}
