% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRTG3D.R
\name{reproduce.track.3d}
\alias{reproduce.track.3d}
\title{Reproduce a track with the eRTG3D}
\usage{
reproduce.track.3d(track, n.sim = 1, multicore = FALSE, error = TRUE,
  DEM = NULL, BG = NULL, plot2d = FALSE, plot3d = FALSE,
  filterDeadEnds = TRUE)
}
\arguments{
\item{track}{data.frame with x,y,z coordinates of the original track}

\item{n.sim}{number of simulations that should be done}

\item{multicore}{logical: run calculations on multiple cores?}

\item{error}{logical: add error term to movement in simulation?}

\item{DEM}{a raster containing a digital elevation model, covering the same extent as the track}

\item{BG}{a raster influencing the probabilities.}

\item{plot2d}{logical: plot tracks on 2d plane?}

\item{plot3d}{logical: plot tracks in 3D?}

\item{filterDeadEnds:}{logical: remove tracks (='NULL') that ended in a dead end?}
}
\value{
A list or data.frame containing the simulated track(s) (CERW).
}
\description{
Simulates n tracks with the geometrical properties of the original track,
between the same start and end point.
}
\examples{
reproduce.track.3d(track)
}
