% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfWrapper.R
\name{sf2df.3d}
\alias{sf2df.3d}
\title{Converts a sf data.frame to a normal dataframe}
\usage{
sf2df.3d(track)
}
\arguments{
\item{track}{An object of type 'sf, data.frame'}
}
\value{
A data.frame.
}
\description{
Converts a sf data.frame to a normal dataframe
}
\examples{
sf2df.3d(df)
}
