% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soaringGliding3D.R
\name{sim.cond.modes.3d}
\alias{sim.cond.modes.3d}
\title{Conditioned Empirical Random Walk (CERW) in 3D with different movement modes}
\usage{
sim.cond.modes.3d(locsVec, start = c(0, 0, 0), end = start, a0, g0, dList,
  qList, error = FALSE, glideRatio = 20, MODE, DEM = NULL, BG = NULL)
}
\arguments{
\item{locsVec}{vector of maximum lengths of modes}

\item{start}{numeric vector of length 3 with the coordinates of the start point}

\item{end}{numeric vector of length 3 with the coordinates of the end point}

\item{a0}{initial incoming heading in radian}

\item{g0}{initial incoming gradient/polar angle in radian}

\item{dList}{list of list objects returned by the \link[eRTG3D]{get.densities.3d} function, one for each MODE}

\item{qList}{list of lists with objects returned by the \link[eRTG3D]{qProb.3d} function, one for each MODE}

\item{error}{logical: add random noise to the turn angle, lift angle and step length to account for errors measurements?}

\item{MODE}{raster layer containing the number/index of the mode, which should be used at each location}

\item{DEM}{raster layer containing a digital elevation model, covering the area between start and end point}

\item{BG}{a background raster layer that can be used to inform the choice of steps}
}
\value{
A trajectory in the form of data.frame
}
\description{
Creates a conditioned empirical random walk, with a specific starting and ending point,
geometrically similar to the initial trajectory, by using different modes of movement.
It is important to extract for every mode in the MODE raster layer a corresponding D object \link[eRTG3D]{get.densities.3d}
and the Q probabilities \link[eRTG3D]{qProb.3d}. The number of steps has to be the same for every mode that is used.
}
\examples{
sim.cond.modes.3d(locsVec, start = c(0,0,0), end=start, a0, g0, dList, qList, MODE)
}
