% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing3D.R
\name{sim.crw.3d}
\alias{sim.crw.3d}
\title{Simulation of a three dimensional Correlated Random Walk}
\usage{
sim.crw.3d(nStep, rTurn, rLift, meanStep, start = c(0, 0, 0))
}
\arguments{
\item{nStep}{the number of steps of the simulated trajectory}

\item{rTurn}{the correlation on the turn angle}

\item{rLift}{the correlation of the lift angle}

\item{meanStep}{the mean step length}

\item{start}{a vector of length 3 containing the coordinates of the start point of the trajectory}
}
\value{
A trajectory in the form of data.frame
}
\description{
Simulation of a three dimensional Correlated Random Walk
}
\examples{
sim.crw.3d(nStep=10, rTurn=0.9, rLift=0.9, meanStep=1, start = c(0,0,0))
}
