% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core3D.R
\name{test.eRTG.3d}
\alias{test.eRTG.3d}
\title{Test the functionality of the eRTG3D}
\usage{
test.eRTG.3d(multicore = FALSE, returnResult = FALSE, plot2d = FALSE,
  plot3d = FALSE)
}
\arguments{
\item{multicore}{logical: test with multicore?}

\item{returnResult}{logical: return tracks generated?}

\item{plot2d}{logical: plot tracks on 2d plane?}

\item{plot3d}{logical: plot tracks in 3D?}
}
\value{
A list containing the original CRW and the simulated track (CERW).
}
\description{
The test simulates a CRW with given parameters and reconstructs it by using the eRTG3D
}
\examples{
test.eRTG3D.3d()
}
