% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing3D.R
\name{test.verification.3d}
\alias{test.verification.3d}
\title{Internally verification of the simulated track}
\usage{
test.verification.3d(track1, track2, alpha = 0.05, plot = FALSE,
  test = "ks")
}
\arguments{
\item{track1}{data.frame or list of data.frames with x,y,z coordinates of the original track}

\item{track2}{data.frame or list of data.frames with x,y,z coordinates of the simulated track}

\item{alpha}{scalar: significance level, default alpha = 0.05}

\item{plot}{logical: plot the densities or differences of turn angle, lift angle and step length of the two tracks?}
}
\value{
Test objects of the 6 two-sample Kolmogorov-Smirnov test conducted.
}
\description{
Uses two-sample Kolmogorov-Smirnov test to compare the geometric characteristics of the original track
with the characteristics of the simulated track.
}
\examples{
test.verification.3d(track1, track2)
}
