% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{track.extent}
\alias{track.extent}
\title{Extent of track(s)}
\usage{
track.extent(track, zAxis = FALSE)
}
\arguments{
\item{track}{a list containing data.frames with x,y,z coordinates or a data.frame}

\item{zAxis}{logical: return also the extent of the Z axis?}
}
\value{
Returns an extent object of the raster package in the 2–D case and a vector in the 3–D case.
}
\description{
Extent of track(s)
}
\examples{
track.extent(niclas, zAxis = TRUE)
}
