% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{track.properties.3d}
\alias{track.properties.3d}
\title{Track properties of a 3-D track}
\usage{
track.properties.3d(track)
}
\arguments{
\item{track}{data.frame with x,y,z coordinates}
}
\value{
The data.frame with track properties
}
\description{
Returns the properties (distances, azimuth, polar angle,
turn angle & lift angle) of a track in three dimensions.
}
\examples{
track.properties.3d(niclas)
}
