% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfWrapper3D.R
\name{track2sf.3d}
\alias{track2sf.3d}
\title{Converts a track to a \code{'sf, data.frame'}}
\usage{
track2sf.3d(track, CRS = NA)
}
\arguments{
\item{track}{eRTG3D track data.frame or a matrix}

\item{CRS}{numeric, EPSG code of the CRS}
}
\value{
A track of type \code{'sf, data.frame'}.
}
\description{
Converts a track to a \code{'sf, data.frame'}
}
\examples{
track2sf.3d(niclas, 4326)
}
