% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfWrapper.R
\name{transformCRS.3d}
\alias{transformCRS.3d}
\title{Transform coordinates reference system of a 3D track}
\usage{
transformCRS.3d(track, fromCRS, toCRS)
}
\arguments{
\item{track}{data.frame with x,y,z coordinates}

\item{fromCRS}{string: proj4 of current CRS}

\item{toCRS}{string: proj4 of CRS to be converted in}
}
\value{
A data.frame containing x,y,z and variables.
}
\description{
Attention: Please use this function for CRS transformations, 
because it is based on the 'st_transform()' from the sf package. Therefore
is supports CRS transdormations in 3D. Note: 'spTransform()' from the 'sp' 
only supports transformations in the 2D plane, which will cause distortions
in the third dimension.
}
\examples{
transformCRS.3d(track, fromCRS="+init=epsg:4326", toCRS="+init=epsg:2056")
}
