% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{turn2target.3d}
\alias{turn2target.3d}
\title{Turn angle to target}
\usage{
turn2target.3d(track)
}
\arguments{
\item{track}{a track data.frame containing x, y and z coordinates}
}
\value{
A numeric vector with the turn angles to target
}
\description{
Calculates the turn angle between every point in the track and the last point (target).
}
\examples{
turn2target.3d(niclas)
}
