% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core3D.R
\name{TurnLiftStepHist}
\alias{TurnLiftStepHist}
\title{3 dimensional histogram}
\usage{
TurnLiftStepHist(turn, lift, step, printDims = TRUE, rm.zeros = TRUE,
  maxBin = 25)
}
\arguments{
\item{turn}{numeric vector of turn angles}

\item{lift}{numeric vector of lift angles}

\item{step}{numeric vector of step lengths}

\item{printDims}{logical: Should dimensions of tld-Cube be messaged?}

\item{rm.zeros}{logical: should combinations with zero probability be removed?}

\item{maxBin}{numeric scalar}
}
\value{
A 3 dimensional histogram as data.frame
}
\description{
Derives a 3 dimensional distribution of a turn angle,
lift angle and step length, by using the Freedman–Diaconis rule for
estimating the number of bins.
}
\examples{
TurnLiftStepHist(turn, lift, step)
}
