% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCloudAnalysis3D.R
\name{chiMaps}
\alias{chiMaps}
\title{Chi maps of two variables}
\usage{
chiMaps(stack1, stack2 = NULL, verbose = FALSE)
}
\arguments{
\item{stack1}{\code{rasterStack}}

\item{stack2}{\code{rasterStack}  \code{NULL} or containing the same number of \code{rasterLayer}s and has euqal extent and resolution.}

\item{verbose}{logical: print currently processed height band in raster stack?}
}
\value{
A \code{rasterStack} containing the chi maps.
}
\description{
Calculates the chi maps for one \code{rasterStack} or all raster all the raster pairs stored in two \code{rasterStack}s.
As observed values, the first stack is used. The expected value is either set to the mean of the first stack, or if given
to be the values of the second stack.
}
\examples{
print("tbd.")
}
