% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfWrapper3D.R
\name{is.sf.3d}
\alias{is.sf.3d}
\title{Tests if the object is a simple feature collection (class: \code{'sf, data.frame'})}
\usage{
is.sf.3d(track)
}
\arguments{
\item{track}{any object to test}
}
\value{
A logical: \code{TRUE} if is a simple feature collection (class: \code{'sf, data.frame'}) of the sf package, \code{FALSE} otherwise.
}
\description{
Tests if the object is a simple feature collection (class: \code{'sf, data.frame'})
}
\examples{
is.sf.3d(niclas)
is.sf.3d(track2sf.3d(track = niclas, CRS = 2056))
}
