% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper3D.R
\name{movingMedian}
\alias{movingMedian}
\title{Moving median in one dimension}
\usage{
movingMedian(data, window)
}
\arguments{
\item{data}{numeric vector}

\item{window}{uneven number for the size of the moving window}
}
\value{
A numeric vector.
}
\description{
Applies a twosided moving median window on a vector,
where the window paramter is the total size of the window.
The value in the window middle is the index where the median of the window is written.
Therefore the window size has to be an uneven number.
The border region of the vetor is filled with a one-sided median.
There might be border effects.
}
\examples{
movingMedian(sequence(1:10), window = 5)
}
