% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCloudAnalysis3D.R
\name{plotRaster}
\alias{plotRaster}
\title{Plots a rasterLayer or rasterStack}
\usage{
plotRaster(r, title = character(0), centerColorBar = FALSE, ncol = NULL)
}
\arguments{
\item{r}{\code{rasterLayer} or \code{rasterStack}}

\item{title}{title text of plot(s)}

\item{centerColorBar}{logical: center colobar around 0 and use \code{RdBuTheme()}?}

\item{ncol}{number of columns to plot a stack, by default estimated by the square root}
}
\value{
Plots the rasters
}
\description{
Plots a rasterLayer or rasterStack
}
\examples{
plotRaster(dem)
}
