% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing3D.R
\name{test.eRTG.3d}
\alias{test.eRTG.3d}
\title{Test the functionality of the eRTG3D}
\usage{
test.eRTG.3d(
  parallel = FALSE,
  returnResult = FALSE,
  plot2d = FALSE,
  plot3d = TRUE,
  plotDensities = TRUE
)
}
\arguments{
\item{parallel}{logical: test running parallel?}

\item{returnResult}{logical: return tracks generated?}

\item{plot2d}{logical: plot tracks on 2-D plane?}

\item{plot3d}{logical: plot tracks in 3-D?}

\item{plotDensities}{logical: plot densities of turning angle, lift angle and step length?}
}
\value{
A list containing the original CRW and the simulated track (CERW).
}
\description{
The test simulates a CRW with given parameters and reconstructs it by using the eRTG3D
}
\examples{
\donttest{
test.eRTG.3d()
}
}
